refactor: move test helper files
