+++ src/table/__tests__/TableHeadWrapper.spec.jsx
@@ -1,4 +1,4 @@
-import '../../../test-setup';
+import './rtl-setup';
 import React from 'react';
 import { render } from '@testing-library/react';
 import { expect } from 'chai';
